requirements.txt
# OG_meeseeks 评估系统依赖包

# 核心依赖
requests>=2.25.0
jsonschema>=4.0.0
json-repair>=0.44.0
python-dotenv>=1.0.0
openai>=2.0.0

# 通用工具
numpy>=1.19.0,<2.0

# 中文处理依赖
pypinyin>=0.44.0
opencc>=1.1.0
hanziconv>=0.3.2

# 日文处理依赖
pykakasi>=2.2.0
janome>=0.4.0
jaconv>=0.3.0

# 韩文处理依赖
hgtk>=0.1.3

# 阿拉伯语处理依赖
PyArabic>=0.6.0
Naftawayh>=0.4

# 德语处理依赖
HanTa>=1.1.0

# 法语处理依赖
verbecc>=1.9.0

# 俄语处理依赖
pymorphy2>=0.9.0
russtress>=0.1.0

# 多语言处理
simplemma>=1.0.0
lingua-language-detector>=2.0.0  # 注意：不是 lingua 包！Python 3.9 使用 2.0.2，3.10+ 使用 2.1+

# 其他工具
Pyphen>=0.10.0
pyspellchecker>=0.7.0
frhyme>=0.3
strokes>=0.0.1
pinyin_order>=0.0.1
haspirater>=0.2

transformers==4.43.4  # 限制版本以兼容 camel_tools

# 注意：以下包有版本冲突但项目需要使用
camel_tools>=1.5.0  # 阿拉伯语高级处理，要求 transformers<4.44.0
char_similar>=0.0.2  # 中文相似字处理（special_chn.py 使用），要求 tqdm==4.31.1
# 警告：char_similar 和 macropodus 要求旧版本依赖，pip 会显示警告但不影响使用